cli/requirements.txt
-e ../sdk
click
requests
pyyaml
